Add resolve module
Resolve creates a symbol table and resolved all symbol references
to either locals (represented by string) or globals (represented by
node id).

I fixed some of the issues in mod_tree too, namely reporting errors
for paths that are yet to be solved (together with duplicate errors)
and issues with „super” glob imports. There are still some bugs that
will get fixed in next commits, because this one is already big enough.

Both resolve and symtable are up to refactoring once I start
implementing the typechecker, when I will be able to see what is
needed in the architecture. Because of that I don't include much
documentation which is, obviously, intended.